requirements.txt
# Core Flask
Flask==3.1.0
Flask-CORS==5.0.0
Flask-RESTX==1.3.0
python-dotenv==1.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# Authentication
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1

# Validation
marshmallow==3.23.1

# Production Server
gunicorn==23.0.0

# Development
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1

# Environment
python-dotenv==1.0.1
